functions/archiemcp/requirements.txt
Flask==3.0.3
google-cloud-aiplatform>=1.49.0